Refactor: 컴포넌트가 아닌 것들 다시 pages 폴더로 이동
Remove: 쓸데없는 코드 제거

Fix: router 설정 잘못되어 있어 404 페이지 뜨는 버그 수정
