+++ src/components/toss/Success.jsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import { useState } from "react";
 import { useSearchParams } from "react-router-dom";
 
-export function SuccessPage() {
+const SuccessPage = () => {
   const [isConfirmed, setIsConfirmed] = useState(false);
   const [searchParams] = useSearchParams();
   const paymentKey = searchParams.get("paymentKey");
@@ -117,4 +117,6 @@ export function SuccessPage() {
       )}
     </div>
   );
-}
+};
+
+export default SuccessPage;
