+++ src/components/toss/index.ts
@@ -1,0 +1,3 @@
+export { default as CheckoutPage } from "./Checkout";
+export { default as FailPage } from "./Fail";
+export { default as SuccessPage } from "./Success";
